Adding support for Import/Export in Base64 text files
