+++ mathics/autoload/formats/Base64/Export.m
@@ -6,18 +6,19 @@ Options[B64Export] = {
     "CharacterEncoding" :> $CharacterEncoding
 };
 
-B64Export[strm_, expr_, OptionsPattern[]] :=
-  Module[{data},
+B64Export[filename_, expr_, OptionsPattern[]] :=
+  Module[{strm, data},
+    strm = OpenWrite[filename];
     If[strm === $Failed, Return[$Failed]];
     data = B64Encode[expr];
-    WriteString[strm, data];
+    WriteString[strm, data];    
     Close[strm];
   ]
 
 ImportExport`RegisterExport[
     "Base64",
 	System`Convert`B64Dump`B64Export,
-	FunctionChannels -> {"Streams"},
+	FunctionChannels -> {"FileNames"},
 	Options -> {"CharacterEncoding", "ByteOrderMark"},
 	DefaultElement -> "Plaintext",
 	BinaryFormat -> True
