+++ mathics/autoload/formats/Base64/Import.m
@@ -5,23 +5,23 @@ Options[B64Import] = {
     "CharacterEncoding" :> $CharacterEncoding
 };
 
-B64Import[strm_, expr_, OptionsPattern[]] := 
-    Module[{data, grid, sep = OptionValue["FieldSeparators"]},
-        data = Read[stream, Record, RecordSeparator->{}];
+B64Import[filename_, OptionsPattern[]] := 
+    Module[{strm,data, grid},
+	strm = OpenRead[filename];
+        data = Read[strm, Record, RecordSeparator->{}];
+	Close[strm];
         { "Data" -> B64Decode[data] }
 ]
 
 ImportExport`RegisterImport[
     "Base64",
-    System`Convert`B64Dump`B64Import,
-    {  },
+    B64Import,
     (* Sources -> ImportExport`DefaultSources["Table"], *)
     FunctionChannels -> {"FileNames"},
     AvailableElements -> {"Data"},
     DefaultElement -> "Data",
     Options -> {
         "CharacterEncoding",
-        "FieldSeparators"
     }
 ]
