+++ mathics/builtin/importexport.py
@@ -528,7 +528,6 @@ class Import(Builtin):
 
     def apply_elements(self, filename, elements, evaluation, options={}):
         'Import[filename_, elements_List?(AllTrue[#, NotOptionQ]&), OptionsPattern[]]'
-
         # Check filename
         path = filename.to_python()
         if not (isinstance(path, six.string_types) and path[0] == path[-1] == '"'):
@@ -542,6 +541,7 @@ class Import(Builtin):
 
         # Load local file
         findfile = Expression('FindFile', filename).evaluate(evaluation)
+
         if findfile == Symbol('$Failed'):
             evaluation.message('Import', 'nffil')
             return findfile
@@ -562,6 +562,7 @@ class Import(Builtin):
 
         for el in elements:
             if not isinstance(el, String):
+
                 evaluation.message('Import', 'noelem', el)
                 return Symbol('$Failed')
 
@@ -587,9 +588,11 @@ class Import(Builtin):
             importer_options.get("System`Options"), options, evaluation)
 
         function_channels = importer_options.get("System`FunctionChannels")
+
         if function_channels is None:
             # TODO message
             return Symbol('$Failed')
+
 
         default_element = importer_options.get("System`DefaultElement")
         if default_element is None:
@@ -981,3 +984,63 @@ class FileFormat(Builtin):
             return None
 
         return from_python(result)
+
+import base64
+
+class B64Encode(Builtin):
+    """
+    <dl>
+    <dt> 'System`Convert`B64Dump`B64Encode[$expr$]'
+    <dd>Encodes $expr$ in Base64 coding
+    </dl>
+
+    >> System`Convert`B64Dump`B64Encode["Hello world"]
+     = IkhlbGxvIHdvcmxkIg==
+    >> System`Convert`B64Dump`B64Decode[%]
+     = Hello world
+    >> System`Convert`B64Dump`B64Encode[Integrate[f[x],{x,0,2}]]
+     = SW50ZWdyYXRlW2ZbeF0sIHt4LCAwLCAyfV0=
+    >> System`Convert`B64Dump`B64Decode[%]
+     = Integrate[f[x], {x, 0, 2}]
+    >> System`Convert`B64Dump`B64Encode[RandomImage[0.,{10,10}]]
+     = ...
+    """
+    
+    context = "System`Convert`B64Dump`"
+    name = "B64Encode"
+
+    def apply(self, expr, evaluation):
+        'System`Convert`B64Dump`B64Encode[expr_]'
+        if isinstance(expr,String):
+            stringtocodify = "\"" + expr.get_string_value() + "\""
+        else:
+            stringtocodify = Expression('ToString',expr).evaluate(evaluation).get_string_value()
+        return String(base64.b64encode(bytearray(stringtocodify, 'utf8')).decode('utf8'))
+            
+
+
+
+class B64Decode(Builtin):
+    """
+    <dl>
+    <dt> 'System`Convert`B64Dump`B64Decode[$string$]'
+    <dd>Decode  $string$ in Base64 coding to an expression.
+    </dl>
+
+    """
+    
+    context = "System`Convert`B64Dump`"
+    name = "B64Decode"
+
+    def apply(self, expr, evaluation):
+        'System`Convert`B64Dump`B64Decode[expr_String]'
+        try:
+            clearstring = base64.b64decode(bytearray(expr.get_string_value(), 'utf8')).decode('utf8')
+        except Exception as e:
+            print(e.__repr__())
+            return Symbol("$Failed")
+        return Expression('ToExpression', String(str(clearstring,))).evaluate(evaluation)
+
+
+
+
